Добавлены экраны подключения, вью модели для них, ресурсы, библиотека и репозиторий
