+++ app/src/main/java/com/dz/bmstu_trade/addDeviceViewModels/DeviceNamingViewModel.kt
@@ -1,4 +1,4 @@
-package com.dz.bmstu_trade.AddDevice
+package com.dz.bmstu_trade.addDeviceViewModels
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
