+++ app/src/main/java/com/dz/bmstu_trade/ui/auth/signin/SignInScreen.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 
+
 @Composable
 fun SignInScreen(navController: NavHostController, onSignIn: () -> Unit) {
     Column(
