+++ app/src/main/java/com/dz/bmstu_trade/ui/main/devicemanualconnect/DeviceConnectionProgressScreen.kt
@@ -1,0 +1,91 @@
+package com.dz.bmstu_trade.ui.main.devicemanualconnect
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.sp
+import com.dz.bmstu_trade.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ConnectionProgressScreen() {
+    Column (
+        modifier = Modifier
+            .fillMaxSize()
+    ) {
+        TopAppBar(
+            title = { Text(text = stringResource(R.string.connection_progress_top_bar_title)) },
+            colors = TopAppBarDefaults.smallTopAppBarColors(
+                containerColor = MaterialTheme.colorScheme.secondaryContainer,
+                titleContentColor = MaterialTheme.colorScheme.onSurface,
+            ),
+            navigationIcon = {
+                IconButton(onClick = { /*TODO*/ }) {
+                    Icon(
+                        imageVector = Icons.Default.ArrowBack,
+                        contentDescription = null
+                    )
+                }
+            }
+        )
+        Box(
+            modifier = Modifier
+                .fillMaxSize(),
+            contentAlignment = Alignment.Center
+        ) {
+            Column {
+                LoadingSpinner()
+                Box (
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = dimensionResource(R.dimen.space_between_spinner_and_label)),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Text(
+                        text = stringResource(R.string.connection_progress_label),
+                        fontSize = dimensionResource(R.dimen.spinner_loader_label_font_size).value.sp
+                    )
+                }
+            }
+        }
+    }
+}
+
+
+@Composable
+fun LoadingSpinner() {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        CircularProgressIndicator(
+            color = Color.Blue,
+            strokeWidth = dimensionResource(R.dimen.spinner_stroke_width),
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun ConnectionProgressPreview() {
+    ConnectionProgressScreen()
+}
